+++ CHANGELOG.md
@@ -9,8 +9,13 @@
 - fix: Umbrella header #671
 - feat: Bump minimum iOS version to 9.0 #669
 - feat: Remove deprecated SDK inits #673
+
+## 5.2.1 
+
+- fix: Add IP address to user serialization #665
 - fix: Crash in SentryEnvelope.initWithEvent #643
 - fix: Build failure for Apple Silicon Macs #588
+- feat: capture userinfo from NSError and NSException #679
 
 ## 5.2.0